Pull up PruneRedundantSortClauses
